Machine Learning and Anomaly Detection/requirements.txt
# Ahşap Anomali Tespiti Projesi - Gereksinimler

# Temel paketler
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
opencv-python>=4.5.0
pillow>=8.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.8.0
torch>=1.10.0
torchvision>=0.11.0

# Görselleştirme ve analiz
jupyter>=1.0.0
ipykernel>=6.0.0

# Görüntü işleme ve veri önişleme
scikit-image>=0.19.0
tqdm>=4.62.0
